backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
supabase>=2.0.0
pydantic>=2.0.0
